Added migration page + responsive and separated components, added layout
